Feat: Add updatePassword to auth but missing the correct URL. Just guessing at the moment.
Refactor: Added some major separation in the UserProfilePage and UserProfile components. All functionality and props will now be on UserProfilePage and be passed down to the UserProfile component. Cleaning up the functions to simplify them and renaming many for ease of reference. UserProfile is now a stateless functional component. Will also likely separate into multiple components.