Refactor: reduce arrows dimension on become-a
It's been reported an overflow on this section of the home.
Unfortunately such scenario has not been encountered in development, the
layout is implemented using css flex features which do not overflow the
parent content unless the parent itself has a width which is bigger than
the container. This needs further review to make sure it won't affect
other devices which couldn't be tested.

Final fix: #58