Refactor: handle menu visibility
This commit handles user menu visibility as follows: when user hovers on
his/her name the menu is visible, when the user enters the menu and then
leaves it the menu is not visible, if the user clicks on an item on the
menu, the menu is not visible. One last behavior is the one that leaves
the menu visible after the user hovers on his/her name, so leaving that
area will still leave the menu visible, this removes a "flickering"
effect while maintaining a quick menu disappearing.